tests: Add exponential retry `backoff library` to authentication tests
We want to be able to retry failing requests before crashing
completely.

Initially, the request library retries 0 times and as such we
are forced to use our `utils.retry` module which makes it possible
to retry a request.

This commit adds the `backoff` library which makes it super easy
to retry failing blocks using decorators in an exponential way.
This library also adds the ability to introduce delays between failing retries
and to determine if an exception is retryable.